NextGen/requirements.txt
flask
flask-cors
psycopg2-binary
gunicorn
pandas
numpy
xgboost
mlxtend
scikit-learn
